# [Core] Optimize `HashMap`  operator`[]` and `insert` method