--- a/core/templates/hash_map.h
+++ b/core/templates/hash_map.h
@@ -68,21 +68,23 @@ template <typename TKey, typename TValue,
 		typename Allocator = DefaultTypedAllocator<HashMapElement<TKey, TValue>>>
 class HashMap {
 public:
-	static constexpr uint32_t MIN_CAPACITY_INDEX = 2; // Use a prime.
-	static constexpr float MAX_OCCUPANCY = 0.75;
+	// Must be a power of two.
+	static constexpr uint32_t INITIAL_CAPACITY = 16;
 	static constexpr uint32_t EMPTY_HASH = 0;
+	static_assert(EMPTY_HASH == 0, "EMPTY_HASH must always be 0 for the memcpy() optimization.");
 
 private:
 	Allocator element_alloc;
-	HashMapElement<TKey, TValue> **elements = nullptr;
-	uint32_t *hashes = nullptr;
 	HashMapElement<TKey, TValue> *head_element = nullptr;
 	HashMapElement<TKey, TValue> *tail_element = nullptr;
+	HashMapElement<TKey, TValue> **elements = nullptr;
+	uint32_t *hashes = nullptr;
 
-	uint32_t capacity_index = 0;
+	// Due to optimization, this is `capacity - 1`. Use + 1 to get normal capacity.
+	uint32_t capacity = 0;
 	uint32_t num_elements = 0;
 
-	_FORCE_INLINE_ uint32_t _hash(const TKey &p_key) const {
+	uint32_t _hash(const TKey &p_key) const {
 		uint32_t hash = Hasher::hash(p_key);
 
 		if (unlikely(hash == EMPTY_HASH)) {
@@ -92,163 +94,174 @@ private:
 		return hash;
 	}
 
-	static _FORCE_INLINE_ uint32_t _get_probe_length(const uint32_t p_pos, const uint32_t p_hash, const uint32_t p_capacity, const uint64_t p_capacity_inv) {
-		const uint32_t original_pos = fastmod(p_hash, p_capacity_inv, p_capacity);
-		return fastmod(p_pos - original_pos + p_capacity, p_capacity_inv, p_capacity);
+	static _FORCE_INLINE_ uint32_t _get_resize_count(uint32_t p_capacity) {
+		return p_capacity ^ (p_capacity + 1) >> 2; // = get_capacity() * 0.75 - 1; Works only if p_capacity = 2^n - 1.
+	}
+
+	static _FORCE_INLINE_ uint32_t _get_probe_length(uint32_t p_pos, uint32_t p_hash, uint32_t p_local_capacity) {
+		const uint32_t original_pos = p_hash & p_local_capacity;
+		return (p_pos - original_pos + p_local_capacity + 1) & p_local_capacity;
 	}
 
 	bool _lookup_pos(const TKey &p_key, uint32_t &r_pos) const {
-		if (elements == nullptr || num_elements == 0) {
-			return false; // Failed lookups, no elements
+		if (unlikely(elements == nullptr)) {
+			return false;
+		}
+		return _lookup_pos_with_hash(p_key, r_pos, _hash(p_key));
+	}
+
+	bool _lookup_pos_with_hash(const TKey &p_key, uint32_t &r_pos, uint32_t p_hash) const {
+		if (unlikely(elements == nullptr)) {
+			return false;
+		}
+		const uint32_t local_capacity = capacity;
+		uint32_t pos = p_hash & local_capacity;
+
+		if (hashes[pos] == p_hash && Comparator::compare(elements[pos]->data.key, p_key)) {
+			r_pos = pos;
+			return true;
 		}
 
-		const uint32_t capacity = hash_table_size_primes[capacity_index];
-		const uint64_t capacity_inv = hash_table_size_primes_inv[capacity_index];
-		uint32_t hash = _hash(p_key);
-		uint32_t pos = fastmod(hash, capacity_inv, capacity);
-		uint32_t distance = 0;
+		if (hashes[pos] == EMPTY_HASH) {
+			return false;
+		}
 
+		// A collision occurred.
+		pos = (pos + 1) & local_capacity;
+		uint32_t distance = 1;
 		while (true) {
-			if (hashes[pos] == EMPTY_HASH) {
-				return false;
-			}
-
-			if (distance > _get_probe_length(pos, hashes[pos], capacity, capacity_inv)) {
-				return false;
-			}
-
-			if (hashes[pos] == hash && Comparator::compare(elements[pos]->data.key, p_key)) {
+			if (hashes[pos] == p_hash && Comparator::compare(elements[pos]->data.key, p_key)) {
 				r_pos = pos;
 				return true;
 			}
 
-			pos = fastmod((pos + 1), capacity_inv, capacity);
+			if (hashes[pos] == EMPTY_HASH) {
+				return false;
+			}
+
+			if (distance > _get_probe_length(pos, hashes[pos], local_capacity)) {
+				return false;
+			}
+
+			pos = (pos + 1) & local_capacity;
 			distance++;
 		}
 	}
 
-	void _insert_with_hash(uint32_t p_hash, HashMapElement<TKey, TValue> *p_value) {
-		const uint32_t capacity = hash_table_size_primes[capacity_index];
-		const uint64_t capacity_inv = hash_table_size_primes_inv[capacity_index];
+	void _insert_with_hash_and_element(uint32_t p_hash, HashMapElement<TKey, TValue> *p_value) {
+		const uint32_t local_capacity = capacity;
+		uint32_t pos = p_hash & local_capacity;
+
+		if (hashes[pos] == EMPTY_HASH) {
+			elements[pos] = p_value;
+			hashes[pos] = p_hash;
+			num_elements++;
+			return;
+		}
+
 		uint32_t hash = p_hash;
 		HashMapElement<TKey, TValue> *value = p_value;
-		uint32_t distance = 0;
-		uint32_t pos = fastmod(hash, capacity_inv, capacity);
-
+		uint32_t distance = 1;
+		pos = (pos + 1) & local_capacity;
 		while (true) {
 			if (hashes[pos] == EMPTY_HASH) {
 				elements[pos] = value;
 				hashes[pos] = hash;
-
 				num_elements++;
-
+#ifdef DEV_ENABLED
+				if (unlikely(distance > 12)) {
+					WARN_PRINT("Excessive collision count (" +
+							itos(distance) + "), is the right hash function being used?");
+				}
+#endif
 				return;
 			}
 
 			// Not an empty slot, let's check the probing length of the existing one.
-			uint32_t existing_probe_len = _get_probe_length(pos, hashes[pos], capacity, capacity_inv);
+			uint32_t existing_probe_len = _get_probe_length(pos, hashes[pos], local_capacity);
 			if (existing_probe_len < distance) {
 				SWAP(hash, hashes[pos]);
 				SWAP(value, elements[pos]);
 				distance = existing_probe_len;
 			}
 
-			pos = fastmod((pos + 1), capacity_inv, capacity);
+			pos = (pos + 1) & local_capacity;
 			distance++;
 		}
 	}
 
-	void _resize_and_rehash(uint32_t p_new_capacity_index) {
-		uint32_t old_capacity = hash_table_size_primes[capacity_index];
+	void _resize_and_rehash(uint32_t p_new_capacity) {
+		const uint32_t real_old_capacity = capacity + 1;
 
-		// Capacity can't be 0.
-		capacity_index = MAX((uint32_t)MIN_CAPACITY_INDEX, p_new_capacity_index);
-
-		uint32_t capacity = hash_table_size_primes[capacity_index];
+		// Capacity can't be 0 and must be 2^n - 1.
+		capacity = MAX(4u, p_new_capacity);
+		const uint32_t real_capacity = next_power_of_2(capacity);
+		capacity = real_capacity - 1;
 
 		HashMapElement<TKey, TValue> **old_elements = elements;
 		uint32_t *old_hashes = hashes;
 
-		num_elements = 0;
-		hashes = reinterpret_cast<uint32_t *>(Memory::alloc_static(sizeof(uint32_t) * capacity));
-		elements = reinterpret_cast<HashMapElement<TKey, TValue> **>(Memory::alloc_static(sizeof(HashMapElement<TKey, TValue> *) * capacity));
+		hashes = reinterpret_cast<uint32_t *>(Memory::alloc_static(sizeof(uint32_t) * real_capacity));
+		elements = reinterpret_cast<HashMapElement<TKey, TValue> **>(Memory::alloc_static(sizeof(HashMapElement<TKey, TValue> *) * real_capacity));
 
-		for (uint32_t i = 0; i < capacity; i++) {
-			hashes[i] = 0;
-			elements[i] = nullptr;
-		}
+		memset(hashes, EMPTY_HASH, real_capacity * sizeof(uint32_t));
 
-		if (old_capacity == 0) {
+		if (old_elements == nullptr) {
 			// Nothing to do.
 			return;
 		}
 
-		for (uint32_t i = 0; i < old_capacity; i++) {
-			if (old_hashes[i] == EMPTY_HASH) {
-				continue;
-			}
+		if (num_elements != 0) {
+			num_elements = 0;
 
-			_insert_with_hash(old_hashes[i], old_elements[i]);
+			for (uint32_t i = 0; i < real_old_capacity; i++) {
+				if (old_hashes[i] == EMPTY_HASH) {
+					continue;
+				}
+
+				_insert_with_hash_and_element(old_hashes[i], old_elements[i]);
+			}
 		}
 
 		Memory::free_static(old_elements);
 		Memory::free_static(old_hashes);
 	}
 
-	_FORCE_INLINE_ HashMapElement<TKey, TValue> *_insert(const TKey &p_key, const TValue &p_value, bool p_front_insert = false) {
-		uint32_t capacity = hash_table_size_primes[capacity_index];
+	// This method does not check that the key is in the table.
+	// Use it if you know for sure that the key is new.
+	HashMapElement<TKey, TValue> *_insert_with_hash(const TKey &p_key, const TValue &p_value, uint32_t p_hash, bool p_front_insert = false) {
 		if (unlikely(elements == nullptr)) {
-			// Allocate on demand to save memory.
-
-			hashes = reinterpret_cast<uint32_t *>(Memory::alloc_static(sizeof(uint32_t) * capacity));
-			elements = reinterpret_cast<HashMapElement<TKey, TValue> **>(Memory::alloc_static(sizeof(HashMapElement<TKey, TValue> *) * capacity));
-
-			for (uint32_t i = 0; i < capacity; i++) {
-				hashes[i] = EMPTY_HASH;
-				elements[i] = nullptr;
-			}
+			_resize_and_rehash(capacity);
+		} else if (unlikely(num_elements > _get_resize_count(capacity))) {
+			_resize_and_rehash(capacity * 2);
 		}
 
-		uint32_t pos = 0;
-		bool exists = _lookup_pos(p_key, pos);
+		HashMapElement<TKey, TValue> *elem = element_alloc.new_allocation(HashMapElement<TKey, TValue>(p_key, p_value));
 
-		if (exists) {
-			elements[pos]->data.value = p_value;
-			return elements[pos];
+		if (tail_element == nullptr) {
+			head_element = elem;
+			tail_element = elem;
+		} else if (p_front_insert) {
+			head_element->prev = elem;
+			elem->next = head_element;
+			head_element = elem;
 		} else {
-			if (num_elements + 1 > MAX_OCCUPANCY * capacity) {
-				ERR_FAIL_COND_V_MSG(capacity_index + 1 == HASH_TABLE_SIZE_MAX, nullptr, "Hash table maximum capacity reached, aborting insertion.");
-				_resize_and_rehash(capacity_index + 1);
-			}
-
-			HashMapElement<TKey, TValue> *elem = element_alloc.new_allocation(HashMapElement<TKey, TValue>(p_key, p_value));
-
-			if (tail_element == nullptr) {
-				head_element = elem;
-				tail_element = elem;
-			} else if (p_front_insert) {
-				head_element->prev = elem;
-				elem->next = head_element;
-				head_element = elem;
-			} else {
-				tail_element->next = elem;
-				elem->prev = tail_element;
-				tail_element = elem;
-			}
-
-			uint32_t hash = _hash(p_key);
-			_insert_with_hash(hash, elem);
-			return elem;
+			tail_element->next = elem;
+			elem->prev = tail_element;
+			tail_element = elem;
 		}
+
+		_insert_with_hash_and_element(p_hash, elem);
+		return elem;
 	}
 
 public:
-	_FORCE_INLINE_ uint32_t get_capacity() const { return hash_table_size_primes[capacity_index]; }
+	_FORCE_INLINE_ uint32_t get_capacity() const { return capacity + 1; }
 	_FORCE_INLINE_ uint32_t size() const { return num_elements; }
 
 	/* Standard Godot Container API */
 
-	bool is_empty() const {
+	_FORCE_INLINE_ bool is_empty() const {
 		return num_elements == 0;
 	}
 
@@ -256,16 +269,14 @@ public:
 		if (elements == nullptr || num_elements == 0) {
 			return;
 		}
-		uint32_t capacity = hash_table_size_primes[capacity_index];
-		for (uint32_t i = 0; i < capacity; i++) {
-			if (hashes[i] == EMPTY_HASH) {
-				continue;
-			}
 
-			hashes[i] = EMPTY_HASH;
-			element_alloc.delete_allocation(elements[i]);
-			elements[i] = nullptr;
+		HashMapElement<TKey, TValue> *current = tail_element;
+		while (current != nullptr) {
+			HashMapElement<TKey, TValue> *prev = current->prev;
+			element_alloc.delete_allocation(current);
+			current = prev;
 		}
+		memset(hashes, EMPTY_HASH, (capacity + 1) * sizeof(uint32_t));
 
 		tail_element = nullptr;
 		head_element = nullptr;
@@ -347,7 +358,7 @@ public:
 		return nullptr;
 	}
 
-	_FORCE_INLINE_ bool has(const TKey &p_key) const {
+	bool has(const TKey &p_key) const {
 		uint32_t _pos = 0;
 		return _lookup_pos(p_key, _pos);
 	}
@@ -360,14 +371,13 @@ public:
 			return false;
 		}
 
-		const uint32_t capacity = hash_table_size_primes[capacity_index];
-		const uint64_t capacity_inv = hash_table_size_primes_inv[capacity_index];
-		uint32_t next_pos = fastmod((pos + 1), capacity_inv, capacity);
-		while (hashes[next_pos] != EMPTY_HASH && _get_probe_length(next_pos, hashes[next_pos], capacity, capacity_inv) != 0) {
+		const uint32_t local_capacity = capacity;
+		uint32_t next_pos = (pos + 1) & local_capacity;
+		while (hashes[next_pos] != EMPTY_HASH && _get_probe_length(next_pos, hashes[next_pos], local_capacity) != 0) {
 			SWAP(hashes[next_pos], hashes[pos]);
 			SWAP(elements[next_pos], elements[pos]);
 			pos = next_pos;
-			next_pos = fastmod((pos + 1), capacity_inv, capacity);
+			next_pos = (next_pos + 1) & local_capacity;
 		}
 
 		hashes[pos] = EMPTY_HASH;
@@ -389,7 +399,6 @@ public:
 		}
 
 		element_alloc.delete_allocation(elements[pos]);
-		elements[pos] = nullptr;
 
 		num_elements--;
 		return true;
@@ -407,24 +416,22 @@ public:
 		HashMapElement<TKey, TValue> *element = elements[pos];
 
 		// Delete the old entries in hashes and elements.
-		const uint32_t capacity = hash_table_size_primes[capacity_index];
-		const uint64_t capacity_inv = hash_table_size_primes_inv[capacity_index];
-		uint32_t next_pos = fastmod((pos + 1), capacity_inv, capacity);
-		while (hashes[next_pos] != EMPTY_HASH && _get_probe_length(next_pos, hashes[next_pos], capacity, capacity_inv) != 0) {
+		const uint32_t local_capacity = capacity;
+		uint32_t next_pos = (pos + 1) & local_capacity;
+		while (hashes[next_pos] != EMPTY_HASH && _get_probe_length(next_pos, hashes[next_pos], local_capacity) != 0) {
 			SWAP(hashes[next_pos], hashes[pos]);
 			SWAP(elements[next_pos], elements[pos]);
 			pos = next_pos;
-			next_pos = fastmod((pos + 1), capacity_inv, capacity);
+			next_pos = (next_pos + 1) & local_capacity;
 		}
 		hashes[pos] = EMPTY_HASH;
-		elements[pos] = nullptr;
-		// _insert_with_hash will increment this again.
+		// _insert_with_hash_and_element will increment this again.
 		num_elements--;
 
 		// Update the HashMapElement with the new key and reinsert it.
 		const_cast<TKey &>(element->data.key) = p_new_key;
 		uint32_t hash = _hash(p_new_key);
-		_insert_with_hash(hash, element);
+		_insert_with_hash_and_element(hash, element);
 
 		return true;
 	}
@@ -432,22 +439,16 @@ public:
 	// Reserves space for a number of elements, useful to avoid many resizes and rehashes.
 	// If adding a known (possibly large) number of elements at once, must be larger than old capacity.
 	void reserve(uint32_t p_new_capacity) {
-		uint32_t new_index = capacity_index;
-
-		while (hash_table_size_primes[new_index] < p_new_capacity) {
-			ERR_FAIL_COND_MSG(new_index + 1 == (uint32_t)HASH_TABLE_SIZE_MAX, nullptr);
-			new_index++;
-		}
-
-		if (new_index == capacity_index) {
+		if (p_new_capacity <= get_capacity()) {
 			return;
 		}
 
-		if (elements == nullptr) {
-			capacity_index = new_index;
-			return; // Unallocated yet.
+		if (unlikely(elements == nullptr)) {
+			capacity = MAX(4u, p_new_capacity);
+			capacity = next_power_of_2(capacity) - 1;
+			return;
 		}
-		_resize_and_rehash(new_index);
+		_resize_and_rehash(p_new_capacity);
 	}
 
 	/** Iterator API **/
@@ -538,7 +539,7 @@ public:
 		return Iterator(tail_element);
 	}
 
-	_FORCE_INLINE_ Iterator find(const TKey &p_key) {
+	Iterator find(const TKey &p_key) {
 		uint32_t pos = 0;
 		bool exists = _lookup_pos(p_key, pos);
 		if (!exists) {
@@ -563,7 +564,7 @@ public:
 		return ConstIterator(tail_element);
 	}
 
-	_FORCE_INLINE_ ConstIterator find(const TKey &p_key) const {
+	ConstIterator find(const TKey &p_key) const {
 		uint32_t pos = 0;
 		bool exists = _lookup_pos(p_key, pos);
 		if (!exists) {
@@ -583,60 +584,73 @@ public:
 
 	TValue &operator[](const TKey &p_key) {
 		uint32_t pos = 0;
-		bool exists = _lookup_pos(p_key, pos);
-		if (!exists) {
-			return _insert(p_key, TValue())->data.value;
-		} else {
+		const uint32_t hash = _hash(p_key);
+		bool exists = _lookup_pos_with_hash(p_key, pos, hash);
+		if (exists) {
 			return elements[pos]->data.value;
+		} else {
+			return _insert_with_hash(p_key, TValue(), hash)->data.value;
 		}
 	}
 
 	/* Insert */
 
 	Iterator insert(const TKey &p_key, const TValue &p_value, bool p_front_insert = false) {
-		return Iterator(_insert(p_key, p_value, p_front_insert));
+		uint32_t pos = 0;
+		const uint32_t hash = _hash(p_key);
+		bool exists = _lookup_pos_with_hash(p_key, pos, hash);
+		if (!exists) {
+			return Iterator(_insert_with_hash(p_key, p_value, hash, p_front_insert));
+		} else {
+			elements[pos]->data.value = p_value;
+			return Iterator(elements[pos]);
+		}
 	}
 
 	/* Constructors */
 
 	HashMap(const HashMap &p_other) {
-		reserve(hash_table_size_primes[p_other.capacity_index]);
-
-		if (p_other.num_elements == 0) {
+		capacity = p_other.capacity;
+		if (p_other.num_elements == 0 || p_other.elements == nullptr) {
 			return;
 		}
 
+		_resize_and_rehash(p_other.capacity);
+
 		for (const KeyValue<TKey, TValue> &E : p_other) {
-			insert(E.key, E.value);
+			uint32_t hash = _hash(E.key);
+			_insert_with_hash(E.key, E.value, hash);
 		}
 	}
 
 	void operator=(const HashMap &p_other) {
 		if (this == &p_other) {
-			return; // Ignore self assignment.
+			return;
 		}
 		if (num_elements != 0) {
 			clear();
 		}
 
-		reserve(hash_table_size_primes[p_other.capacity_index]);
-
 		if (p_other.elements == nullptr) {
-			return; // Nothing to copy.
+			return;
+		}
+
+		if (p_other.capacity > capacity) {
+			_resize_and_rehash(p_other.capacity);
 		}
 
 		for (const KeyValue<TKey, TValue> &E : p_other) {
-			insert(E.key, E.value);
+			uint32_t hash = _hash(E.key);
+			_insert_with_hash(E.key, E.value, hash);
 		}
 	}
 
 	HashMap(uint32_t p_initial_capacity) {
-		// Capacity can't be 0.
-		capacity_index = 0;
-		reserve(p_initial_capacity);
+		capacity = MAX(4u, p_initial_capacity);
+		capacity = next_power_of_2(capacity) - 1;
 	}
-	HashMap() {
-		capacity_index = MIN_CAPACITY_INDEX;
+	HashMap() :
+			capacity(INITIAL_CAPACITY - 1) {
 	}
 
 	HashMap(std::initializer_list<KeyValue<TKey, TValue>> p_init) {
@@ -646,27 +660,31 @@ public:
 		}
 	}
 
-	uint32_t debug_get_hash(uint32_t p_index) {
-		if (num_elements == 0) {
-			return 0;
-		}
-		ERR_FAIL_INDEX_V(p_index, get_capacity(), 0);
-		return hashes[p_index];
-	}
-	Iterator debug_get_element(uint32_t p_index) {
-		if (num_elements == 0) {
-			return Iterator();
-		}
-		ERR_FAIL_INDEX_V(p_index, get_capacity(), Iterator());
-		return Iterator(elements[p_index]);
-	}
-
 	~HashMap() {
-		clear();
-
 		if (elements != nullptr) {
+			HashMapElement<TKey, TValue> *current = tail_element;
+			while (current != nullptr) {
+				HashMapElement<TKey, TValue> *prev = current->prev;
+				element_alloc.delete_allocation(current);
+				current = prev;
+			}
+
 			Memory::free_static(elements);
 			Memory::free_static(hashes);
+			elements = nullptr;
+			hashes = nullptr;
+			tail_element = nullptr;
+			head_element = nullptr;
+			num_elements = 0;
+			capacity = INITIAL_CAPACITY - 1;
 		}
 	}
 };
+
+extern template class HashMap<int, int>;
+extern template class HashMap<String, int>;
+extern template class HashMap<StringName, bool>;
+extern template class HashMap<StringName, StringName>;
+extern template class HashMap<StringName, String>;
+extern template class HashMap<StringName, Variant>;
+extern template class HashMap<StringName, int>;