… core/templates/hashfuncs.cpp
… core/templates/hashfuncs.h
… tests/core/templates/test_hash_map.h