**Benchmarks**

Compile command: `scons -j11 target=template_release scu_build=true scu_limit=70 tests=yes`

**HashMap<String, int>**
| Time Before| Time After|
|-----------------|---------------|
| 53299 | 50615 |
| 55542 | 49761 |
| 55685 | 49801 |
| 56682 | 51170 |
| 53138 | 49786 |
| 58661 | 55219 |
| 56247 | 54025 |
| 54254 | 49417 |

Avarage `before - after`:  4214.25
Avarage % improved: 7.60166671176168

![image](https://github.com/godotengine/godot/assets/116561933/a27de104-321c-4353-a9df-8ad6666c7df6)

**HashMap<int, int>**

| Time before | Time after |
|-------------|------------|
| 12097       | 10191      |
| 11364       | 10436      |
| 12383       | 10497      |
| 12094       | 10230      |
| 11048       | 10122      |
| 11416       | 10183      |
| 10794       | 10468      |
| 11952       | 10250      |

Avarage `before - after`: 1346.375
Avarage % improved: 11.5633185897711

![image](https://github.com/godotengine/godot/assets/116561933/6ad3808f-1e75-4802-aa4a-9fee91066ba0)