@Nazarwadim 2024-04-01
Now it looks like this 
![image](https://github.com/godotengine/godot/assets/116561933/169d23b7-3a95-42df-8a6b-1b7d34217281)