@MewPurPur 2024-04-02
What is measured in "before" and "after"?