@Nazarwadim 2024-04-02
> What is measured in "before" and "after"?

Master branch - before.
My branch - after