@Nazarwadim 2024-04-02
> No _what_ is measured, what are the units

u_sec