@Nazarwadim 2024-04-02
benchmark code:
```c++
HashMap1<int, int> before;
HashMap<int, int> after;
int found1;
int found2;
uint32_t count = 100000;
uint32_t find_count = 100000;
void test_hash_map() {
	uint64_t start1 = OS::get_singleton()->get_ticks_usec();
	for (uint32_t i = 0; i < count; i++) {
		before[rand()] = rand();
	}
	uint64_t end1 = OS::get_singleton()->get_ticks_usec();
	uint64_t start2 = OS::get_singleton()->get_ticks_usec();
	for (uint32_t i = 0; i < count; i++) {
		after[rand()] = rand();
	}
	uint64_t end2 = OS::get_singleton()->get_ticks_usec();
	print_line("Time inserting before:", end1 - start1);
	print_line("Time inserting after:", end2 - start2);

	uint64_t start3 = OS::get_singleton()->get_ticks_usec();

	for (uint32_t i = 0; i < find_count; i++) {
		auto e = before.getptr(rand());
		if (e) {
			found1 = *e;
		}
	}

	uint64_t end3 = OS::get_singleton()->get_ticks_usec();
	before.clear();
	uint64_t start4 = OS::get_singleton()->get_ticks_usec();

	for (uint32_t i = 0; i < find_count; i++) {
		auto e = after.getptr(rand());
		if (e) {
			found2 = *e;
		}
	}
	uint64_t end4 = OS::get_singleton()->get_ticks_usec();
	after.clear();
	print_line("Time find before:", end3 - start3);
	print_line("Time find after:", end4 - start4);
}
```