@kleonc 2024-04-02
> benchmark code:
> (...)

Each before/after case should be runned for the same dataset, otherwise the before/after execution times might differ not only because of the changes made in this PR but also because of e.g. different hash collisions for the input datasets.

Like in the example above before/after should be runned with the same RNG seed.