@AThousandShips 2024-04-02
`getptr` getting worse twice, by a non-insignificant amount, is worrying, to me it implies something about these improvements are broken, there's no reason for it to become slower