@Nazarwadim 2024-04-02
> Try restoring `fastmod` and see what changes, I have a feeling that is the cause, I don't think it can be simply removed

`fastmod` and my version work exactly the same.  I did benchmarks and my version is 2 times faster. Also, it doesn't degrade by half, but only by about 5%. As you can see, after I started using `RandomPCG`, the insertion has degraded by half and the search by 3. I think the best thing to do would be to try to create some random array and then insert its elements into the table and see the results. And then go through the array looking for those elements from the table.