@AThousandShips 2024-04-02
> Also, it doesn't degrade by half, but only by about 5%

I know? I didn't say it did? I said it was a reduction, and it doesnt make sense if the input is exactly the same