@Nazarwadim 2024-04-02
> I know? I didn't say it did? I said it was a reduction, and it doesnt make sense if the input is exactly the same

2 times faster