@Nazarwadim 2024-04-02
> I don't even see how the replacement code is the same, do I don't think that's a valid change

For example
``` c++
capacity = 10;
hash = 15;
pos = hash % capacity; // first fastmod now  pos = 5 

//then in the loop we do something like this

pos = (pos + 1) % capacity // 6 % 10 = 6
// after a few iterations pos = 9;
pos = (pos + 1) % capacity // 10 % 10 = 0
```