@AThousandShips 2024-04-02
You're still looking up at an infinitesimal chance of hitting