@Nazarwadim 2024-04-02
![image](https://github.com/godotengine/godot/assets/116561933/00b5db66-fa9a-4728-a1d7-5ef265489cc1)
![image](https://github.com/godotengine/godot/assets/116561933/59a43c02-968e-41df-a78e-22ebcaba8c9f)