@Nazarwadim 2024-04-03
> 1. Now we're optimizing, wouldn't (`un`)`likely()` help in some checks? I'm specifically talking about the various `exists`/`!exists`.

It makes no difference.

>     2. Regarding `fastmod()`:
        
`pos + 1 < capacity ? (pos + 1) : 0;` and `fastmod` are not the same. 
fastmod = %.
My variation is to make `pos` go around in a circle.
Something like this
1 2 3 4 5 6 7 8 9 0 1 2 3 .......