@Nazarwadim 2024-04-03
I found that it is possible to completely remove `fastmod` in the `_get_probe_length` method