@MewPurPur 2024-04-03
@Geometror I think you implemented fastmod, maybe you can provide feedback?