@Geometror 2024-04-30
Alright, checking this again and to be honest I'm not sure about completely replacing fastmod with &, since that requires having a capacity of 2^n. That might work for synthetic benchmarks, but generally there is a good reason for using prime numbers for the capacity (as many other hash map implementations do as well): to prevent clustering and achieve a better distribution, thus preventing a high number of collisions (that can even be a problem with very good hash functions).
As far as I know you're only testing with completely random data, but that's not really representative of the real-world usage. For example, (in case I remember correctly) TileSet/TileMap makes extensive use of HashMap, using integer coordinates as keys.
The other optimizations you did however look fine to me :)