@Nazarwadim 2024-04-30
The problem about & seems to exist only when the hash does not differ on the lower bits. `hash_fmix32` seems to solve this problem.
JDK uses 2^n and has a similar function. 
https://hg.openjdk.org/jdk7/jdk7/jdk/file/9b8c96f96a0f/src/share/classes/java/util/HashMap.java#l264